Add Avatar component
A new Avatar component was created and is now being exported from the 'src/components/ui/avatar' folder. This component can take in a 'size' prop, which can either be 'large' or 'small', and additional properties meant for the <img> element. If a 'src' prop is provided, the component will display the image provided, otherwise it displays a default icon.