Added Header component and Logo component
Implemented a new Header component and a Logo component. The Header component includes user info display and navigation links. When the user is logged in, their nickname and avatar will be displayed; otherwise, a Sign In button will be shown. The Logo component is a new SVG element included in the header.