Add 'createNewPasswordForm' component
New 'createNewPasswordForm' component has been created to allow users to create a new password. The form, based on the react-hook-form and zod schema for field validation, includes a controlled text field and a clickable button to initiate the password creation process. Overall, this addition facilitates new password creation and further instructions are set to be sent via email.